vcd2step: Converts VCD files to Chisel tester inputs
It turns out that VCD files contain all the information required to
drive a Chisel tester.  By combining a VCD with a Flo file it is
possible to reconstruct which values should be poked in as inputs and
then generate the cooresponding Chisel tester file.

The purpose of this program is to do exactly that.