Initialize all registers in the target design
The way that Chisel computs things means that VCD time 0 actually
contains the register values after the first cycle of initialization.
This means that it's impossible to actually get the values directly
from the VCD file.

Luckily, we have the flo file that produces this circuit.  That means
I can do a graph traversal in order to discover the initialization
value of registers.  Unfortunately this only works when they match the
pattern.